+++ src/Gaa.DotNet.Templates/content/gaa-project-library/LogicLayer.cs
@@ -1,0 +1,17 @@
+using System.Diagnostics.CodeAnalysis;
+
+namespace Gaa.Project.Library;
+
+/// <summary>
+/// Маркерный класс для сборки.
+/// </summary>
+[ExcludeFromCodeCoverage]
+public sealed class LibraryLayer
+{
+    /// <summary>
+    /// Инициализирует новый экземпляр класса <see cref="LibraryLayer"/>.
+    /// </summary>
+    private LibraryLayer()
+    {
+    }
+}